refactor(user): move class to appropriate component
